checkpoint 1 of moving bootstrap
