+++ xds/src/main/java/io/grpc/xds/ClusterResolverLoadBalancer.java
@@ -52,9 +52,13 @@ import io.grpc.xds.EnvoyServerProtoData.SuccessRateEjection;
 import io.grpc.xds.EnvoyServerProtoData.UpstreamTlsContext;
 import io.grpc.xds.PriorityLoadBalancerProvider.PriorityLbConfig;
 import io.grpc.xds.PriorityLoadBalancerProvider.PriorityLbConfig.PriorityChildConfig;
-import io.grpc.xds.XdsClient.ResourceWatcher;
-import io.grpc.xds.XdsEndpointResource.EdsUpdate;
-import io.grpc.xds.XdsLogger.XdsLogLevel;
+import io.grpc.xds.client.Locality;
+import io.grpc.xds.client.XdsClient;
+import io.grpc.xds.client.XdsClient.ResourceWatcher;
+import io.grpc.xds.client.XdsEndpointResource;
+import io.grpc.xds.client.XdsEndpointResource.EdsUpdate;
+import io.grpc.xds.client.XdsLogger;
+import io.grpc.xds.client.XdsLogger.XdsLogLevel;
 import java.net.URI;
 import java.net.URISyntaxException;
 import java.util.ArrayList;
