+++ xds/src/main/java/io/grpc/xds/CsdsService.java
@@ -33,10 +33,13 @@ import io.grpc.Status;
 import io.grpc.StatusException;
 import io.grpc.internal.ObjectPool;
 import io.grpc.stub.StreamObserver;
-import io.grpc.xds.XdsClient.ResourceMetadata;
-import io.grpc.xds.XdsClient.ResourceMetadata.ResourceMetadataStatus;
-import io.grpc.xds.XdsClient.ResourceMetadata.UpdateFailureState;
-import io.grpc.xds.XdsNameResolverProvider.XdsClientPoolFactory;
+import io.grpc.xds.client.SharedXdsClientPoolProvider;
+import io.grpc.xds.client.XdsClient;
+import io.grpc.xds.client.XdsClient.ResourceMetadata;
+import io.grpc.xds.client.XdsClient.ResourceMetadata.ResourceMetadataStatus;
+import io.grpc.xds.client.XdsClient.ResourceMetadata.UpdateFailureState;
+import io.grpc.xds.client.XdsClientPoolFactory;
+import io.grpc.xds.client.XdsResourceType;
 import java.util.Map;
 import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
