+++ xds/src/main/java/io/grpc/xds/EnvoyServerProtoData.java
@@ -22,6 +22,7 @@ import com.google.common.collect.ImmutableList;
 import com.google.protobuf.util.Durations;
 import io.envoyproxy.envoy.extensions.transport_sockets.tls.v3.CommonTlsContext;
 import io.grpc.Internal;
+import io.grpc.xds.client.EnvoyProtoData;
 import io.grpc.xds.internal.security.SslContextProviderSupplier;
 import java.net.InetAddress;
 import java.net.UnknownHostException;
