+++ xds/src/main/java/io/grpc/xds/InternalSharedXdsClientPoolProvider.java
@@ -18,6 +18,7 @@ package io.grpc.xds;
 
 import io.grpc.Internal;
 import io.grpc.internal.ObjectPool;
+import io.grpc.xds.client.SharedXdsClientPoolProvider;
 import java.util.Map;
 
 /**
