+++ xds/src/main/java/io/grpc/xds/InternalXdsAttributes.java
@@ -23,6 +23,9 @@ import io.grpc.Internal;
 import io.grpc.NameResolver;
 import io.grpc.internal.ObjectPool;
 import io.grpc.xds.XdsNameResolverProvider.CallCounterProvider;
+import io.grpc.xds.client.FilterChainSelectorManager;
+import io.grpc.xds.client.Locality;
+import io.grpc.xds.client.XdsClient;
 import io.grpc.xds.internal.security.SslContextProviderSupplier;
 
 /**
