+++ xds/src/main/java/io/grpc/xds/client/XdsLogger.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.grpc.xds;
+package io.grpc.xds.client;
 
 import com.google.common.base.Preconditions;
 import io.grpc.InternalLogId;
