+++ xds/src/test/java/io/grpc/xds/ClusterResolverLoadBalancerTest.java
@@ -76,7 +76,11 @@ import io.grpc.xds.PriorityLoadBalancerProvider.PriorityLbConfig;
 import io.grpc.xds.PriorityLoadBalancerProvider.PriorityLbConfig.PriorityChildConfig;
 import io.grpc.xds.RingHashLoadBalancer.RingHashConfig;
 import io.grpc.xds.WrrLocalityLoadBalancer.WrrLocalityConfig;
-import io.grpc.xds.XdsEndpointResource.EdsUpdate;
+import io.grpc.xds.client.Locality;
+import io.grpc.xds.client.XdsEndpointResource;
+import io.grpc.xds.client.XdsEndpointResource.EdsUpdate;
+import io.grpc.xds.client.XdsClient;
+import io.grpc.xds.client.XdsResourceType;
 import io.grpc.xds.internal.security.CommonTlsContextTestsUtil;
 import java.net.SocketAddress;
 import java.net.URI;
