+++ xds/src/test/java/io/grpc/xds/CsdsServiceTest.java
@@ -46,11 +46,17 @@ import io.grpc.internal.ObjectPool;
 import io.grpc.internal.testing.StreamRecorder;
 import io.grpc.stub.StreamObserver;
 import io.grpc.testing.GrpcServerRule;
-import io.grpc.xds.Bootstrapper.BootstrapInfo;
-import io.grpc.xds.Bootstrapper.ServerInfo;
-import io.grpc.xds.XdsClient.ResourceMetadata;
-import io.grpc.xds.XdsClient.ResourceMetadata.ResourceMetadataStatus;
-import io.grpc.xds.XdsNameResolverProvider.XdsClientPoolFactory;
+import io.grpc.xds.client.Bootstrapper.BootstrapInfo;
+import io.grpc.xds.client.Bootstrapper.ServerInfo;
+import io.grpc.xds.client.EnvoyProtoData;
+import io.grpc.xds.client.XdsClient;
+import io.grpc.xds.client.XdsClient.ResourceMetadata;
+import io.grpc.xds.client.XdsClient.ResourceMetadata.ResourceMetadataStatus;
+import io.grpc.xds.client.XdsClientPoolFactory;
+import io.grpc.xds.client.XdsClusterResource;
+import io.grpc.xds.client.XdsEndpointResource;
+import io.grpc.xds.client.XdsListenerResource;
+import io.grpc.xds.client.XdsResourceType;
 import java.util.Collection;
 import java.util.HashMap;
 import java.util.List;
