+++ xds/src/test/java/io/grpc/xds/EnvoyProtoDataTest.java
@@ -21,8 +21,10 @@ import static com.google.common.truth.Truth.assertThat;
 import com.google.common.collect.ImmutableMap;
 import com.google.protobuf.Struct;
 import com.google.protobuf.Value;
-import io.grpc.xds.EnvoyProtoData.Address;
-import io.grpc.xds.EnvoyProtoData.Node;
+import io.grpc.xds.client.EnvoyProtoData;
+import io.grpc.xds.client.EnvoyProtoData.Address;
+import io.grpc.xds.client.EnvoyProtoData.Node;
+import io.grpc.xds.client.Locality;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
