+++ xds/src/test/java/io/grpc/xds/FilterChainSelectorManagerTest.java
@@ -22,9 +22,10 @@ import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
 import io.grpc.ServerInterceptor;
 import io.grpc.xds.EnvoyServerProtoData.FilterChain;
-import io.grpc.xds.FilterChainMatchingProtocolNegotiators.FilterChainMatchingHandler.FilterChainSelector;
-import io.grpc.xds.FilterChainSelectorManager.Closer;
-import io.grpc.xds.XdsServerWrapper.ServerRoutingConfig;
+import io.grpc.xds.client.FilterChainMatchingProtocolNegotiators.FilterChainMatchingHandler.FilterChainSelector;
+import io.grpc.xds.client.FilterChainSelectorManager;
+import io.grpc.xds.client.FilterChainSelectorManager.Closer;
+import io.grpc.xds.client.XdsServerWrapper.ServerRoutingConfig;
 import java.util.Collections;
 import java.util.concurrent.atomic.AtomicReference;
 import org.junit.Test;
