+++ xds/src/test/java/io/grpc/xds/LoadReportClientTest.java
@@ -53,6 +53,8 @@ import io.grpc.stub.StreamObserver;
 import io.grpc.testing.GrpcCleanupRule;
 import io.grpc.xds.LoadStatsManager2.ClusterDropStats;
 import io.grpc.xds.LoadStatsManager2.ClusterLocalityStats;
+import io.grpc.xds.client.EnvoyProtoData;
+import io.grpc.xds.client.Locality;
 import java.util.ArrayDeque;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
